Poseiden-skeleton/src/main/java/com/nnk/springboot/web/controllers/HomeController.java
package com.nnk.springboot.web.controllers;

import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;

import javax.annotation.security.RolesAllowed;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.security.Principal;

@Controller
public class HomeController
{
	@RequestMapping("/")
	public String home(Model model)
	{
		return "home";
	}

	@RequestMapping("/user/home")
	public String adminHome()
	{
		return "redirect:/bidList/list";
	}

}
